add parsing process reporting facilities
Add the callback that will notify the parser's caller about what is
happening within the parser. Shift and reduce operations are broadcast.

Right now it can be seen that the final reduction is not reported if
there is no possibility to continue parsing (i.e., if the goal
production terminates with EOF). It means that the reporting will have
to be reworked :)