improve branch id handling; refactorings
First, the shift() method that makes a shifted-by-one stack moved from
State to ItemStack.

Second, a new method was introduced: ItemStack.copyWithNoId(). With it,
State objects gain a new property: no stack remaining in a State after
its construction references a branch ID object that it doesn't need.
This way, we enable users to safely use as branch IDs the objects that
contain the branch state (the incomplete parse tree, etc.).