remove 'symbols' parameter from GlrCallback.shift()
Step 1: realize that there can be either one symbol (usually) or no
symbols (when initializing the very first branch).

Step 2: realize that, when there is a symbol, it's the same for all
invocations within the given invocation of State.advance().

The above means that the symbols parameter is unnecessary: initial vs
usual invocation can be differentiated between by nullness of oldBranch,
and if the callback needs the exact symbol, it's the caller's problem.