use TreeStack in LoggingCallback; simplify State initialization
The driving idea behind this change is to use TreeStack in the
LoggingCallback. As the transition was being made, it became clear that
the use of null as the identifier of the empty stack branch forces us to
adopt the same approach within State. From that, State.initializeFrom()
became simler (does not require the callback anymore).